Enhancement#4: Implementation of HistoryTextHandler
