--- src/main/java/com/ngram/app/DummyHistoryTextHandler.java
@@ -1,21 +1,0 @@
-package com.ngram.app;
-
-import com.ngram.app.browser.NgordnetQuery;
-import com.ngram.app.browser.NgordnetQueryHandler;
-
-import java.util.List;
-
-public class DummyHistoryTextHandler extends NgordnetQueryHandler {
-    @Override
-    public String handle(NgordnetQuery q) {
-        List<String> words = q.words();
-        int startYear = q.startYear();
-        int endYear = q.endYear();
-
-        String response = "You entered the following info into the browser:\n";
-        response += "Words: " + q.words() + "\n";
-        response += "Start Year: " + q.startYear() + "\n";
-        response += "End Year: " + q.endYear() + "\n";
-        return response;
-    }
-}
